19 - Edición y eliminación de registros - Elimnar registros con restricciones de Clave Foránea en Laravel

+++ app/Http/Controllers/ProfessionController.php
@@ -8,12 +8,16 @@ use Illuminate\Http\Request;
 class ProfessionController extends Controller
 {
     public function index(){
-    	return view('profession.index', [
-    		'professions' => Profession::orderBy('title')->get()
-    	]);
+		$professions = Profession::query()
+			->withCount('profiles')
+			->orderBy('title')
+			->get();
+
+    	return view('profession.index', compact('professions'));
     }
 
     public function destroy(Profession $profession){
+		abort_if($profession->profiles()->exists(), 400, 'Cannot delete a profession linked to a profile.');
     	$profession->delete();
     	return redirect(route('profession.index'));
     }
